finishing the basics actions to phone endpoints

+++ OrgManager.API/Controllers/PhoneController.cs
@@ -46,45 +46,141 @@ namespace src.OrgManager.API.Controllers
             }
             catch (Exception ex)
             {
-                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Organização. Erro: {ex.Message}");
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Telefone. Erro: {ex.Message}");
             }
         }
 
-        [HttpPost("Organization/{id}")]
-        public async Task<IActionResult> PostByOrganization(int id, PhoneDto model)
-        {
-            try
-            {
-                var userId = 0;
-                if(await _accountService.UserExistsInOrganization(User.GetUserId(), id) == null) 
-                    return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
-                var phone = await _phoneService.AddPhone(userId, id, model);
-                if(phone == null) return NoContent();
-
-                return Ok(phone);
-            }
-            catch (Exception ex)
-            {
-                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Organização. Erro: {ex.Message}");
-            }
-        }
-
-        [HttpDelete("Organization/{organizationId}/{phoneId}")]
-        public async Task<IActionResult> DeleteByOrganization(int organizationId, int phoneId)
+        [HttpPost("Organization/{organizationId}")]
+        public async Task<IActionResult> PostByOrganization(int organizationId, PhoneDto model)
         {
             try
             {
                 var userId = 0;
                 if(await _accountService.UserExistsInOrganization(User.GetUserId(), organizationId) == null) 
                     return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
-                var phone = await _phoneService.DeletePhone(userId, organizationId, phoneId);
-                if(phone == false) return this.StatusCode(StatusCodes.Status404NotFound, $"Telefone não encontrado para exclusão.");
+                var phone = await _phoneService.AddPhone(userId, organizationId, model);
+                if(phone == null) return NoContent();
 
                 return Ok(phone);
             }
             catch (Exception ex)
             {
-                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Organização. Erro: {ex.Message}");
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpPut("User")]
+        public async Task<IActionResult> PutByUser(PhoneDto model)
+        {
+            try
+            {
+                int organizationId = 0;
+                int userId = User.GetUserId();
+                PhoneDto phone = await _phoneService.UpdatePhone(userId, organizationId, model);
+                if(phone == null) return NoContent();
+
+                return Ok(phone);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar Alterar Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpGet("User/{phoneId}")]
+        public async Task<IActionResult> GetPhone(int phoneId)
+        {
+            try
+            {
+                int organizationId = 0;
+                int userId = User.GetUserId();
+                PhoneDto phone = await _phoneService.GetPhoneByIdAsync(userId, organizationId, phoneId);
+
+                if(phone == null) return NoContent();
+
+                return Ok(phone);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar consultar Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpGet("Organization/{organizationId}/{phoneId}")]
+        public async Task<IActionResult> GetPhone(int organizationId, int phoneId)
+        {
+            try
+            {
+                int userId = 0;
+                if(await _accountService.UserExistsInOrganization(User.GetUserId(), organizationId) == null) 
+                    return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
+                PhoneDto phone = await _phoneService.GetPhoneByIdAsync(userId, organizationId, phoneId);
+
+                if(phone == null) return NoContent();
+
+                return Ok(phone);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar consultar Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpGet("User")]
+        public async Task<IActionResult> GetAllPhones()
+        {
+            try
+            {
+                int organizationId = 0;
+                int userId = User.GetUserId();
+                PhoneDto[] phones = await _phoneService.GetAllPhonesByIdAsync(userId, organizationId);
+
+                if(phones == null) return NoContent();
+
+                return Ok(phones);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar consultar Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpGet("Organization/{organizationId}")]
+        public async Task<IActionResult> GetAllPhones(int organizationId)
+        {
+            try
+            {
+                int userId = 0;
+                if(await _accountService.UserExistsInOrganization(User.GetUserId(), organizationId) == null) 
+                    return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
+                PhoneDto[] phones = await _phoneService.GetAllPhonesByIdAsync(userId, organizationId);
+
+                if(phones == null) return NoContent();
+
+                return Ok(phones);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar consultar Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpPut("Organization/{organizationId}")]
+        public async Task<IActionResult> PutByOrganization(int organizationId, PhoneDto model)
+        {
+            try
+            {
+                int userId = 0;
+                if(await _accountService.UserExistsInOrganization(User.GetUserId(), organizationId) == null) 
+                    return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
+                PhoneDto phone = await _phoneService.UpdatePhone(userId, organizationId, model);
+                if(phone == null) return NoContent();
+
+                return Ok(phone);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar Alterar Telefone. Erro: {ex.Message}");
             }
         }
 
@@ -93,16 +189,35 @@ namespace src.OrgManager.API.Controllers
         {
             try
             {
-                var organizationId = 0;
+                int organizationId = 0;
                 int userId = User.GetUserId();
-                var phone = await _phoneService.DeletePhone(userId, organizationId, phoneId);
+                bool phone = await _phoneService.DeletePhone(userId, organizationId, phoneId);
                 if(phone == false) return this.StatusCode(StatusCodes.Status404NotFound, $"Telefone não encontrado para exclusão.");
 
                 return Ok(phone);
             }
             catch (Exception ex)
             {
-                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar incluir Organização. Erro: {ex.Message}");
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar Excluir Telefone. Erro: {ex.Message}");
+            }
+        }
+
+        [HttpDelete("Organization/{organizationId}/{phoneId}")]
+        public async Task<IActionResult> DeleteByOrganization(int organizationId, int phoneId)
+        {
+            try
+            {
+                int userId = 0;
+                if(await _accountService.UserExistsInOrganization(User.GetUserId(), organizationId) == null) 
+                    return this.StatusCode(StatusCodes.Status400BadRequest, $"Usuário não existe na Organização.");
+                bool phone = await _phoneService.DeletePhone(userId, organizationId, phoneId);
+                if(phone == false) return this.StatusCode(StatusCodes.Status404NotFound, $"Telefone não encontrado para exclusão.");
+
+                return Ok(phone);
+            }
+            catch (Exception ex)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar Excluir Telefone. Erro: {ex.Message}");
             }
         }
     }
